+++ instrumentation/aws-sdk/aws-sdk-2.2/library/src/main/java/io/opentelemetry/instrumentation/awssdk/v2_2/internal/BedrockRuntimeAttributesGetter.java
@@ -27,7 +27,7 @@ class BedrockRuntimeAttributesGetter
   }
 
   @Override
-  public String getProviderName(ExecutionAttributes executionAttributes) {
+  public String getSystem(ExecutionAttributes executionAttributes) {
     return GenAiProviderNameIncubatingValues.AWS_BEDROCK;
   }