+++ instrumentation/openai/openai-java-1.1/library/src/main/java/io/opentelemetry/instrumentation/openai/v1_1/ChatAttributesGetter.java
@@ -27,7 +27,7 @@ final class ChatAttributesGetter
   }
 
   @Override
-  public String getProviderName(ChatCompletionCreateParams request) {
+  public String getSystem(ChatCompletionCreateParams request) {
     return GenAiAttributes.GenAiProviderNameIncubatingValues.OPENAI;
   }