+++ instrumentation/openai/openai-java-1.1/library/src/main/java/io/opentelemetry/instrumentation/openai/v1_1/EmbeddingAttributesGetter.java
@@ -25,7 +25,7 @@ final class EmbeddingAttributesGetter
   }
 
   @Override
-  public String getProviderName(EmbeddingCreateParams request) {
+  public String getSystem(EmbeddingCreateParams request) {
     return GenAiAttributes.GenAiProviderNameIncubatingValues.OPENAI;
   }